+++ R/GetRegData.R
@@ -69,12 +69,12 @@ WHERE
   query <- paste0(query, year, " AND var.AvdRESH IN (", deps, ");")
 
   if ("session" %in% names(list(...))) {
-    raplog::repLogger(session = list(...)[["session"]],
+    rapbase::repLogger(session = list(...)[["session"]],
                       msg = paste("Load tilsynsrapport data from",
                                   registryName, ": ", query))
   }
 
-  rapbase::LoadRegData(registryName, query, dbType)
+  rapbase::loadRegData(registryName, query, dbType)
 }
 
 
@@ -91,7 +91,7 @@ SELECT
   PasientID,
   ForlopsID,
   InklKritOppf,
-  SkrSamtykke
+  SkriftligSamtyk
 FROM
   AlleVarNum
 WHERE
@@ -101,11 +101,11 @@ WHERE
                   startDate, "' AND '", endDate, "');")
 
   if ("session" %in% names(list(...))) {
-    raplog::repLogger(session = list(...)[["session"]],
+    rapbase::repLogger(session = list(...)[["session"]],
                       msg = paste0("Load data from ", registryName, ":", query))
   }
 
-  rapbase::LoadRegData(registryName, query, dbType)
+  rapbase::loadRegData(registryName, query, dbType)
 }
 
 #' @rdname getRegData
@@ -143,7 +143,11 @@ SELECT
   var.StSmBev21,
   var.PasientID,
   var.ForlopsID,
-  var.InnlAvd
+  var.InnlAvd,
+  var.VidereOppf,
+  var.BehNedtrappAvsluttTils,
+  var.Journalnotat,
+  var.IkkeMedBeh
 FROM
   AlleVarNum var
 WHERE
@@ -156,12 +160,12 @@ WHERE
   }
 
   if ("session" %in% names(list(...))) {
-    raplog::repLogger(session = list(...)[["session"]],
+    rapbase::repLogger(session = list(...)[["session"]],
                       msg = paste("Load indikatorrapport data from",
                                   registryName, ": ", query))
   }
 
-  rapbase::LoadRegData(registryName, query, dbType)
+  rapbase::loadRegData(registryName, query, dbType)
 }
 
 #' @rdname getRegData
@@ -183,7 +187,7 @@ GROUP BY
   YEAR(RegDato11);
 ")
 
-  rapbase::LoadRegData(registryName, query, dbType)
+  rapbase::loadRegData(registryName, query, dbType)
 }
 
 #' @rdname getRegData
@@ -201,7 +205,7 @@ GROUP BY
   YEAR(RegDato11);
 ")
 
-  rapbase::LoadRegData(registryName, query, dbType)
+  rapbase::loadRegData(registryName, query, dbType)
 }
 
 
@@ -232,7 +236,7 @@ GROUP BY
   if (reshId %in% conf$reg$smerte$nationalAccess$reshId) {
     return("Nasjonal")
   } else {
-    df <- rapbase::LoadRegData(registryName, dbType = dbType, query = query)
+    df <- rapbase::loadRegData(registryName, dbType = dbType, query = query)
     n <- dim(df)[1]
     hVec <- df[1:n, 1]
     if (n > 1) {
@@ -295,8 +299,8 @@ WHERE
   }
 
   if ("session" %in% names(list(...))) {
-    raplog::repLogger(session = list(...)[["session"]],
+    rapbase::repLogger(session = list(...)[["session"]],
                       msg = paste("Smerte data dump:\n", query))
   }
-  rapbase::LoadRegData(registryName, query)
+  rapbase::loadRegData(registryName, query)
 }